Split user model into more files
