+++ client-v2/src/app/models/auth.model.ts
@@ -1,0 +1,18 @@
+import { UserPreview } from '.'
+import { HttpSuccessResponse } from '../http/types'
+
+export interface SignupCredentialsDto {
+    username: string
+    email: string
+    password: string
+}
+export interface LoginCredentialsDto {
+    email: string
+    password: string
+}
+
+export type LoggedInUser = UserPreview & {
+    authToken: string
+}
+
+export type AuthSuccessResponse = HttpSuccessResponse<{ user: LoggedInUser }>
