+++ client-v2/src/app/models/index.ts
@@ -1,0 +1,4 @@
+export interface UserPreview {
+    id: string
+    username: string
+}
