+++ client-v2/src/app/pages/auth/login/login.component.ts
@@ -4,9 +4,9 @@ import { Actions } from '@ngrx/effects'
 import { Store } from '@ngrx/store'
 import { FormBuilderOptions } from 'src/app/components/molecules/form/types'
 import { betterEmailValidator } from 'src/app/components/molecules/form/validators'
+import { LoginCredentialsDto } from 'src/app/models/auth.model'
 import { AppState } from 'src/app/store'
 import { userActions } from 'src/app/store/user/user.actions'
-import { LoginCredentialsDto } from 'src/app/store/user/user.model'
 import { getErrorMap } from '../getErrorMap'
 
 @Component({
