+++ client-v2/src/app/store/user/user.actions.ts
@@ -1,6 +1,6 @@
 import { createAction, props } from '@ngrx/store'
 import { HttpServerErrorResponse } from 'src/app/http/types'
-import { LoggedInUser, LoginCredentialsDto, SignupCredentialsDto } from './user.model'
+import { SignupCredentialsDto, LoginCredentialsDto, LoggedInUser } from 'src/app/models/auth.model'
 
 export const userActions = {
     signup: createAction('[User] signup', props<SignupCredentialsDto>()),
