+++ client-v2/src/app/store/user/user.model.ts
@@ -1,4 +1,4 @@
-import { HttpSuccessResponse } from 'src/app/http/types'
+import { UserPreview } from 'src/app/models'
 
 export interface UserState {
     me: UserPreview | null
@@ -6,26 +6,3 @@ export interface UserState {
     isLoggedIn: boolean
     isLoading: boolean
 }
-
-// @TODO: These should be in another file
-
-export interface UserPreview {
-    id: string
-    username: string
-}
-
-export interface SignupCredentialsDto {
-    username: string
-    email: string
-    password: string
-}
-export interface LoginCredentialsDto {
-    email: string
-    password: string
-}
-
-export type LoggedInUser = UserPreview & {
-    authToken: string
-}
-
-export type AuthSuccessResponse = HttpSuccessResponse<{ user: LoggedInUser }>
